Move DetailedView to /anime/:id, support direct links
I like a url pattern like /anime/123 for the anime detail page of anime 123.  This is similar to links for content on other services (https://netflix.com/title/70299043, etc), provides clear and distinct links for sharing, and allows users to land directly on an anime detail page and still have it work (previously it would not, since it counted on the anime data to be provided in location state by another page).