Fix possible infinite loop of API calls
Although it doesn't seem possible, if the API somehow returns success for getting an anime, but the data has the wrong anime id, this hook will keep retrying the fetch as soon as the previous one is done.  This happens because it only prevents the API call when the fetchedAnime or cachedAnime id matches the id we are looking for. It is better to defensively handle this situation and treat it like an error if the API returns mismatched anime data.  This won't change any of the useEffect dependencies (animeId, fetchedAnime, cachedAnime), so useEffect won't re-run or trigger another request.

I still don't love this hook.  It is possible to make it hammer the API with identical requests for an anime by just changing the cachedAnime input a bunch.  This is unlikely to happen, but not great design.  The best way to prevent this would probably be to implement an API request caching layer, which is hard, and best left to a library like react-query.  So I will leave it as-is, until we decide if we want to add a fetch-caching solution or roll our own.