Give anime detail pages unique urls (#7)
* Add /anime/{id} GET function to APICalls

* Move DetailedView to /anime/:id, support direct links

I like a url pattern like /anime/123 for the anime detail page of anime 123.  This is similar to links for content on other services (https://netflix.com/title/70299043, etc), provides clear and distinct links for sharing, and allows users to land directly on an anime detail page and still have it work (previously it would not, since it counted on the anime data to be provided in location state by another page).

* Fix possible infinite loop of API calls

Although it doesn't seem possible, if the API somehow returns success for getting an anime, but the data has the wrong anime id, this hook will keep retrying the fetch as soon as the previous one is done.  This happens because it only prevents the API call when the fetchedAnime or cachedAnime id matches the id we are looking for. It is better to defensively handle this situation and treat it like an error if the API returns mismatched anime data.  This won't change any of the useEffect dependencies (animeId, fetchedAnime, cachedAnime), so useEffect won't re-run or trigger another request.

I still don't love this hook.  It is possible to make it hammer the API with identical requests for an anime by just changing the cachedAnime input a bunch.  This is unlikely to happen, but not great design.  The best way to prevent this would probably be to implement an API request caching layer, which is hard, and best left to a library like react-query.  So I will leave it as-is, until we decide if we want to add a fetch-caching solution or roll our own.

* Refactor anime matching logic to avoid duplication

This logic is pretty important to be consistent in the useEffect call and outside it.  So, I extracted it to a helper function to make this obvious.

* User helper fn everywhere

Probably should use my anime-matching helper everywhere that ids are checked.