Call /anime/:id/similar api endpoint on DetailedView
I was finding that the /recommend used on DetailedView was messing up sometimes and not returned what I would expect to see, since it would include all the user's likes/dislikes in the call, which would often steer the similar titles to basically be the same as 'For You' on home.  I think this shelf on the anime's page should just be the anime that are most similar to the title.  Calling /anime/:id/similar will get those, specifically.

Later, we may want to add some personalization to the similar call, so that the similar titles are also re-ranked based on predicted relevance to the user.  But if we want to do that, I will have to adjust the API and recommendation engine -- will probably get most similar titles first, then re-rank those candidates based on a second relevance scoring model.