Fix bug when useSimilarAnime input args change
Somehow I didn't notice at first that if the input values change (animeId or amount), the previous fetched value no longer matches, and this means we need to issue a new request.  This never happened in the uses of this hook so far, but this change fixes the issue anyways.

I've fixed it by generating a unique string key out of the input args (animeId, params) and storing this alongside the fetched results.  I then make sure that, if I have a fetched result, its key matches the one for the current params the user has passed in.  If not, this is an old result, and we need to throw it away and issue a request for the new params.