Fix bug where 'null episodes' can appear on anime page (#10)
If a show has not finished yet, it can be possible for the 'episodes' field to be null.  It is my mistake that I didn't notice this and include it in the API documentation.  In these cases, I just shorten it to print the format.  A good example is the show One Piece (/anime/21).

I also extracted the logic to determine the byline text to a helper function -- although I love a good ternary, in this case I would have had to add a third case and it was getting pretty long.