Add AnimeCard, AnimeGrid
Some slight trickiness here since the AnimeGrid needs to keep track of which AnimeCard child is selected so that it can set its z-index higher than the other grid items in case they overlap when the selected one grows.  I chose a simple callback pattern where the AnimeCard can be given a function to call back when its selected state changes.

Had to raise the header's z-index so it is always over the cards.

Had to change the LikeButton and DislikeButton to prevent the event default when clicked on, since otherwise, the card link would trigger when the user likes things inside the card.