Implement new AnimeCards for onboarding
They have a slightly different interaction style to match the fact that you can only like items, and clicking on them should like them, instead of taking you to the anime page.  I feel like they are distinct from the cards on home, but also match the style of the site.  Let me know what you think.

There was also an extra Container in Onboarding around the text, so I removed it, which makes the text line up with the grid of items nicely.  Hopefully that is ok.  If it was supposed to be there, I can put it back and put the grid inside it.