Use "inherit" as the default LikeButton color
This will match whatever the font color is at its current location.  This is nice because it can be black when on the white background of the page, next to black text, and it can be white when in the anime cards, next to white text on the black gradient overlay.