Demo of using react components as JSX tags
You don't have to do things this way, especially if you prefer calling your custom functional components as functions, but I thought I would send this to just demonstrate how you can do the same thing but use your component as a JSX tag.  Instead of listing a bunch of arguments, you just specify all the prop values you want as an object expansion of the first 'props' argument of the function, then those are available as attributes on the tag you can set.  This method of making/using custom components is pretty common so just wanted to make sure you have seen it and are familiar with it.  If you prefer the other method, that is cool too.