Break apart generic api calls into diff useEffects (#18)
This reduces the number of reloads when the user taps a chip.  This makes the animation when chips get selected much less likely to be interrupted by re-render jank.

There still is the problem of the initial page load where each shelf comes in at a diff time and the page adjusts its layout like 8 times in rapid succession, but that can be fixed in a follow-up change.