Style chips to match mocks
As a part of this, I now am using the Chip `variant` prop to indicate selected state -- a selected chip is filled, while the rest are outlined.  I've added a border and used a grey-ish color for the selected chip.

Since the `color` property of every genre object in GenreChips state is no longer needed, I did remove it and all logic that works with it.  This is what I would called 'derived' state anyways, in that the correct color/variant/appearance can be derived by just knowing the selected state.