Use `await` to simplify acct linking functions
Since these linkWithGoogle/Twitter/etc functions are all async, we can leverage await instead of using the .then().catchError() pattern.  This prevents the need for a lot of nested thens.

There are some situations where you can't make functions async, such as when providing a function argument to useEffect, etc.  In those cases, it is good to be familiar with the Promise.then().catchError() pattern.  But when you can use async/await, it often reads better.

I do have an open question on when we want to call setForwardToken.  If I understand correctly, this may start a navigation back on the registration page to the home page.  In the linkWithEmail function, I notice that it waits till almost all the async operations are done before calling setForwardToken(true).  However, in the Google/Twitter ones, this is called right after linkWithPopup succeeds and we get our credential, but before we update the firebase user and update the document in firestore with the account info.  Which of those do we think is best?  Since we are merging documents, it is probably not broken to have a race between the setDoc in the linking function that updates name/authProvider/etc and the SaveToFirestore in register that saves the like/dislike fields.  No matter which one wins the race, neither overwrites the other.  But might it be safer to just always wait until all writes are done before triggering any transitions back on the registration page?