Add rudimentary AnimeShelf, use on /home, /anime
This shelf allows for horizontal navigation of multiple pages of content, and I've installed it for the generic shelves on Home (except random) and the similar items shelf on DetailedView.

In the future, I'd like this shelf to have a cool scrolling animation.  Right now it just instantly switches to the new page of content.  I'd also like to have it reset to the first page whenever the genre filter changes.

I've tried to make it responsive, in that the number of items onscreen will slowly be reduced as the screen size shrinks.  I did need to add a new breakpoint to bridge an awkward spot between sm and md.