Style DetailedView, including new UrlButtons
I've added some icons for the streaming services that match our phosphorous icon style, though maybe it would be nice for these to instead be colorful.  I will keep messing with it.

I've tried to make the page responsive enough to look acceptable on smaller screens, but I am not totally thrilled with it yet.