Remove unneeded navigate call on Home
There is a navigation to the page we are already on when the user is logged-in, which will likely have no effect, so I have removed it.

I also got rid of a console.log from here, which helped in my debugging to keep superfluous prints out of the console.  In general, when we get ready to productionize this app, console.logs should be avoided, or used only in cases where we want the user to see the information while helping us debug.