Remove Login's guest button
It doesn't make sense to be able to login as a guest, I don't think.  Since the anonymous id is gone forever once the user logs out, doing this won't restore any previous guest sessions, as far as I can tell.  So this is just a way to log in for the first time without going through onboarding.  I think it would probably be better to just not have this option, since authenticating as guest is always pretty much a new registration.