Fix Login as Guest failures (#29)
* Update useLikeState to not fail when LocalUser is empty

This should never happen, but may in a case where the user logs in as someone who has no LocalUser in Firestore (login as guest without going through onboarding can repro this, for example)

* Remove Login's guest button

It doesn't make sense to be able to login as a guest, I don't think.  Since the anonymous id is gone forever once the user logs out, doing this won't restore any previous guest sessions, as far as I can tell.  So this is just a way to log in for the first time without going through onboarding.  I think it would probably be better to just not have this option, since authenticating as guest is always pretty much a new registration.