Fix like/dislike deletion on login (#28)
* Don't overwrite user profile on login

Right now, on login from Login, the user's existing likes and dislikes are overwritten with the temp ones from being in a logged-out context.  It seems like only Registration should involve writing likes/dislikes to firestore on success.  Otherwise, any logging-in user will lose all their likes and dislikes.

* Remove unneeded navigate call on Home

There is a navigation to the page we are already on when the user is logged-in, which will likely have no effect, so I have removed it.

* Wait for logout to succeed before clearing LocalUser

I made logout an async function so we can wait for its success before continuing with the logout process.  We shouldn't clear localUser or navigate to /logout if the logout operation fails.

---------

Co-authored-by: Ken Metz <metzke@umich.edu>