Use staleTime for recommendations api call
Similarly, to fully take advantage of react-query, this change converts useGetRecommendations to stop using passed-in state setters, and instead rely on useQuery's result object.  In Home, the `data` and `isLoading` fields from this result object are used to replace the former useStates for `recommend` and `loadingRecs` state variables.  This means that on repeat visits to the Home page, the recommendations won't be continually re-fetched.  They will be refetched if the user changes their likes or dislikes, since the viewHistory is part of the useQuery cache key.  Right now, this means that the loading spinner re-appears for the 'For You' section every time that the user updates a like or dislike on Home.  A follow-up commit will address this and try to make that smoother.