Use handleErrors util function in new fetchers
This nice helper function will check the response status, and if not ok, throw an error containing the error text from the HTTP response body.  It basically does the same thing as the code it is replacing, with the advantage of being in a shared function for more consistency.  Also, I made sure to check the response status before accessing other field son response, since we don't want to try to parse the response body as json in case of errors.