Adds API call caching using TanStack Query (#31)
* Adds 'TanStack Query' to cache/streamline api calls

* Use staleTime for generic list requests

The requests for the generic anime lists can be cached for a little while since they are not likely to change often.  So if the user changes pages a lot, we won't need to re-request them every time they return to Home.  This more fully leverages the advantage of react-query.  To do this I had to:
1. Set a `staleTime` in the useQuery options.  I've chosen 5 minutes but we can mess with this.  Repeat requests for the same anime list will reuse the cached result during this window.
2. Use the data returned by useQuery.  As it was written, anime list state was actually set in the Home component via passing around the setState setters.  Although clever, this setter was placed inside the fetching function which means that it would only happen once when a fetch was triggered.  Return visits to the page that used a cached useQuery result without triggering the fetcher would never set state on Home.  However, useQuery itself returns a 'data' field which will contain either undefined (if loading or error) or else the last good data from the fetch.  I rewrote functions GetAnimeM* to, first, use hook naming, since they call hooks and therefore are hooks, and second, return the useQuery result object.  Then Home could access this data, whether from a fetch or from cache.  Home itself will re-render whenever the return from those useQuery calls changes, which happens on request completion.

* Use staleTime for recommendations api call

Similarly, to fully take advantage of react-query, this change converts useGetRecommendations to stop using passed-in state setters, and instead rely on useQuery's result object.  In Home, the `data` and `isLoading` fields from this result object are used to replace the former useStates for `recommend` and `loadingRecs` state variables.  This means that on repeat visits to the Home page, the recommendations won't be continually re-fetched.  They will be refetched if the user changes their likes or dislikes, since the viewHistory is part of the useQuery cache key.  Right now, this means that the loading spinner re-appears for the 'For You' section every time that the user updates a like or dislike on Home.  A follow-up commit will address this and try to make that smoother.

* Use handleErrors util function in new fetchers

---------

Co-authored-by: Steve Metz <metzsg@gmail.com>