Add loading states to AnimeGrid/Shelf when no items (#32)
Right now, this just involves showing "ghost" cards in the grid or shelf.  It is still not perfect since when you like items in For You, it replaces them all with ghost cards very temporarily until the new recs come in, whereas I almost liked it more when the old cards stayed there until we got the new recs.  Maybe another loading indicator can be used.  I will address in a follow-up, but I want to get this in since it solves jankiness now that we use react-query.