Add ghosts cards for onboarding items
This will make it a little nicer if the user has to wait for the anime to load in for onboarding.

Eventually, I think it would be a good idea to curate the anime shown in onboarding to represent a range of types of content (lots of action titles there right now), and maybe to extend the list to be 6-per-row, which would make the cards shorter so the "Let's go" button would be above the fold.