Use AnimeShelf for 'For You' to allow more recs
This commit extends AnimeShelf to be able to have multiple rows, and uses it on Home for the "For You" items.

This matches the behavior of all the other shelves on the site.  Maybe weird, since I don't know how common it is for horizontal shelves to have two rows, but also maybe the most similar to every other UI interaction and so the most natural.