Add 'see more' button to 'For You' section
This commit adds a button to the bottom of the "For You" grid that shows more content in the grid.  This allows for a vertical scroll to view more personalized recs, which might be more fluid on mobile especially.  It also prevents the need for a paradigm where a shelf can have multiple rows, which is somewhat unusual, I think.  This approach is maybe analogous to the shorts items on youtube home, which extend down if you click a 'more' button.

There is an alternate option in branch 'feature/for-you-shelf' which basically creates a two-row horizontal shelf that is similar to other shelves on Home.