Persist appSettings to localStorage
This will allow the user's darkMode preference to survive page reloads, and return visits to the page in a new tab, etc.

As a part of this work, I also separated `AppSettingsContext` and its provider component into two separate files.  This is because React hot-reloads will break contexts if you hot-reload files that execute `createContext`, since this creates a brand new context object on hot-reload, while dependent child components are still looking for a provider of the *old* context object, which it can't find.  So the value becomes undefined and any component that does `useContext()` on it will break.  This shouldn't affect real users, but is annoying during dev if you are working on the provider.