Implement dark mode (#45)
* Implement dark mode

Dark mode works by:
- Creating an `AppSettingsContext` that stores the user's selected mode.
- Creating a `AppThemeProvider` which uses the AppSettingsContext to regenerate a new theme when this mode changes, and put this in a MUI `ThemeProvider` that wraps the rest of the app.
- Adding mui's `CssBaseline` component so that body bg and text default colors come from mui's theme.
- Adjusting all components to no longer refer to explicit colors, and instead pull them from the theme, *unless they are in a place where the colors don't change due to theme, like LandingPage's hero display, etc*.
- Adding a new version of the logout Spike image that has a transparent background, new versions of streaming service icons in white.

I've put the dark mode UI control in `DropMenu` for now.

From now on, when adding new UI, colors that will change with the mode should come from the material theme.  If the theme doesn't have a color, we can add more.  Some basics to know:
- `theme.palette.text.primary` is the main text color
- `theme.palette.background.default` is the main bg color
- `theme.palette.primary.main` is the brand red color.

https://mui.com/material-ui/customization/dark-mode/ has a list of all the mode-dependent palette colors.

If you make sure to import `useTheme` from `@mui/material`, VSCode's autocomplete can help you find your way around the theme.  The generated theme has many more fields than are set in the call to `createTheme` in `theme.js`.  There are some other packages that also export `useTheme` without the cool type information, so be sure to check which package it is coming from.  I've changed all `useTheme` imports that I could find to be from the right spot.

A known issue is that the dark mode settings don't persist if you reload the page or navigate directly to a url via the browser.  I will solve that in a follow-up.

* Persist appSettings to localStorage

This will allow the user's darkMode preference to survive page reloads, and return visits to the page in a new tab, etc.

As a part of this work, I also separated `AppSettingsContext` and its provider component into two separate files.  This is because React hot-reloads will break contexts if you hot-reload files that execute `createContext`, since this creates a brand new context object on hot-reload, while dependent child components are still looking for a provider of the *old* context object, which it can't find.  So the value becomes undefined and any component that does `useContext()` on it will break.  This shouldn't affect real users, but is annoying during dev if you are working on the provider.