Adjust some header spacing
This change aims to tighten up the spacing between shelf headers and their content, and make pages fit better above the fold.  Previously, the headers were equidistant from their shelf and the previous shelf.  Moving them closer to the shelf they describe feels a bit better, as the association is more clear.  I've reduced the standard 'gap' to only 20px, to reduce the whitespace between the header and the page contents.  I like that this lets the 'see more' button be above the fold (on my 15" macbook, at least)  I also add a standard bottom padding to the html body so that every page will have a little space at the end without explicitly needing a gap div there.  Finally, I reduced the number of characters in DetailedView text so that the descriptions aren't quite as tall in their compacted form.