Fix bugs with undefined bio and avatar fields
This pops up with accounts that pre-date these features.  In the case of the one in Firestore.js, firestore errors out if you try to save documents that have fields set to undefined, so I've used null instead as the empty value for bio and avatar.  Feel free to change that to empty string if you prefer.