Normalize avatar images, update some for quality
I've updated the portrait images to all be 160x160 jpgs.  I tried to select similar pictures to the ones originally added, but in some cases I had to choose a slightly different picture to either get the color temperatures to line up right (the character was from a still in a show where their skin was tined weird colors by the lighting of the scene or something), or to get the quality high enough that all the images look matching.  I couldn't for the life of me find a good portrait shot of rebecca to work with, so I replaced her with another character from the mocks, but we can add her back if we can find a good portrait that plays well with the others.  Not totally set on these as the final images of these characters (Naruto's has some compression artifacts from the source image that I don't love).  In the end, it will probably be ok to have some variance in the image color temps and qualities, as long as "sets" of images from each show all look consistent (i.e. Naruto's characters may never look as HD as those from a new show, but if the row of Naruto avatars all match well, and the row of each other show's avatars match, then it should work)