Fix bug where AnimeShelf causes mobile pages to overflow
AnimeShelf overflows on the x-axis once Container's side padding downgrades to only 16px on small screen widths, causing a horizontal scrollbar and other unpleasant stuff mobile browsers do when the page doesn't fit the screen width.

This fix does cause a very slight height shift in the shelf when leaving the first page, which I don't love, but I absolutely don't love the buttons being partly offscreen or horizontal scrollbars for the page.

Using the AnimeShelf on a phone, it is pretty obvious the best UX would be a swipe-to-scroll shelf, so I will work on that.