Header tab styling (#58)
* Remove extraneous 'onKeyDown'

* Make Header tab icon color text.primary

This will match all the other icons, and the header text when on large screens.

* Remove unused useEffect

* Refactor header tabs into HeaderTab component

Will be useful for next change that introduces stylish underlines under the active tab

* Add underline beneath active tab

As seen in the mocks.  Lmk if you like this or not -- there are other, more subtle ways we can indicate the selected tab

* Center header tabs a bit more in `md`, `lg`

Achieved by making the title and drop menu grid items equal in size, so the center grid has an equal buffer on either side.  I also got rid of some `lg` specifications since they were equal to `md` ones and therefore not needed.

* Fix onKeyDown for search that I broke

Plus also I got rid of the `navigate` call since the Link component should handle the navigation for us

* Fix issue with hovering/clicking search icon

My bad!  The icon should also be a click target and should change the HeaderTab's color when hovering.

* Nudge HeaderTab underline up

Just messing around with it.  This is what I did in the mocks before I started overthinking everything