Add ability to play anime trailers in a modal
I've added the button to open the trailer modal on top of the cover on DetailedView, like rottentomatoes has.  The trailer videos coming from the API are not ideal right now, as a lot of them are in Japanese with no subtitles.  Not sure if we want to hold off on adding this until our trailers are better -- that is another option, and we can leave this unmerged until then.