Fix bug where loading indicator shows too much (#65)
The loading indicator on the "Data From Edward" section of DetailedView should only show if it is loading data for a new anime.  Updating analysis of an anime shouldn't show the indicator.  I renamed a field in the /analyze API response which broke this.  I thought it wasn't being used and forgot that I used it.  My bad!