Make the avatar button slightly smaller
Since there is limited space over in the sidebar, I removed some of the padding from the avatar button in ProfileUserBanner and moved the text a little closer to it, so longer usernames can fit without wrapping or overflowing.  I also rounded the corners a little bit, to match the style of the other box-like things in the app.