Enforce default value of `reviews` in LocalUser
Otherwise, with my legacy account that didn't have this field, I got errors while trying to save my profile or open the review authoring form.

I think, in general, this will be a good practice as we add new fields to LocalUser.  The alternative is checking if it is null everywhere that it is read, which is a lot more error-prone since forgetting even one time can cause bugs, and leads to more code everywhere.