Create 'ghost' versions of DetailedView, Profile
These will show while some core data dependencies are loading.  The annoying thing about these is that they do introduce a maintenance burden, since we have to update them if we update the layout of the pages, and they only are seen rarely, usually during cold-load of a url pointing to one of the pages, or one first load of a DetailedView while the analysis loads.  An alternative might be to just use a loading edward on the page.