Add a subtitle to the ProfileListPage
It currently has the type of list ("Watchlist", "Watch History" for likes/dislikes) and the number of items in the list.  I figure it could be extended to have the number of likes or saves, if we let people like or save watchlists, and other stuff like the top genres in the lists or something. Lmk if you like this idea, we don't have to go with it if it feels redundant or clutter-y.