Fix colors and clicking on like buttons (#119)
At head right now, like buttons on AnimeCards don't have a `selected` value passed in, so their icons are all background-colored.  Also, I noticed clicking on the like/dislike buttons on AnimeCards also navigates to the detail page, so I added back the `e.preventDefault()` to their click handlers.  The 'Add' button already has this, so it was still working and didn't need it.