Move 'back' button up from beside title to beside avatar (#122)
* Move 'back' button up from beside title to beside avatar

On ProfileListPage, I don't really like how the back button inset the title and content below it.  Even worse, we had to add extra logic to `ClickAndEdit`, a general component, to support this extra left margin.
 Overall, I like a design where the big titles are always flush with the left side of the content section they describe.  To do this, I:

- Added a new `BackButton` component
- Removed the `ml` prop from `ClickAndEdit`
- Moved the back button up to be beside the `ProfileUserBannerSmall` on profile list pages.

The new `BackButton` component has an `sx` prop that is piped through to the mui `IconButton` inside, so it can be used to add style props just like with a mui component.  Sadly, it doesn't get the cool autocomplete features when you are filling out the style in vscode, though 😢.

* fix typo

---------

Co-authored-by: Ken Metz <metzke@umich.edu>