Add more watch buttons for new sites like Hulu, Disney+
Now that the api serves more watch urls, I've added options to show these. We will also show "Google It" options if a google url is provided from the API with a google mid (machine id, basically a knowledge graph id). And `UrlButtons` will now show a "Google It" button with a naive search query if there are no watch urls present.