Fix some bugs with loading list reactions (#137)
This commit fixes an issue where the ghost chips won't show up many times while the list reactions are loading.  This was because the `useState` provided a default value for listRxns that wouldn't let the component know the data wasn't available yet.  I've solved by using undefined until the data is present.  And, `GetListReactions` will now return a default reaction object when the document doesn't yet exist, which will allow for creating reactions on lists that don't have reactions yet.