Add 'poster' to edward gif for safari
Safari doesn't like the mp4 video (I think it is due to the way the safari browser tries to request it in streaming format, but not sure).  Right now, due to this bug, the gif's spot is blank on the landing page.  To get something to show up instead, I added a 'poster' attribute to the video tag so that a still image would show up there.  In fact, this image will show up for all browsers until the video loads in, so should be a win all around, although this is below the fold, and the video is small, so other browser users probably wouldn't even notice.