Fix error message display
Since the object being thrown is now an Error object (`throw new Error("uh oh!")` instead of a string `throw "uh oh!"`), we will have to check the message property of the error object in the error handler.

I also made sure this error, and the 'unkownClaimError' one, would show up and tested it out.  This way of unpacking the errors one at a time is a little verbose and hard to maintain, but that seems to be how yup does it.  I bet we could write a reducer or something to make this *just work* with new error types in the future, but this will be ok for now