Choose handle (#144)
* Adds user handles:
- Forces prompt on /home for all non-anonymous users w/o handles
- Saves handles to a new Firestore collection
- Implements form validation to prevent duplicate handle names

* Delete old/unused javascript files

* - Normalizes handle names before checking for duplicates
- Fixes loading state in 'RoutingHelper.js' that would get stuck when user logged out.

* Prevent escaping the 'choose a handle' dialog by pressing 'escape' key

* Make handle dialog responsive

* Clean-up unneeded code

* Adds a breathing logo to the submit button on <HandleDialog.js>

* Remove leading @ symbol for users without a handle (ie. anonymous accounts)

* Bug Fix - Bug: User's registering as a guest account got stuck on a loading screen. Cause: Populating the localUser in 'RoutingHelper.js' broke the registration page logic (setForwardToken would never get triggered). Solution: Drilled down the 'HandleDialog.js' component.

* Implement transactions to ensure multiple users can't pull the same handle name

* Reworks registration process for email/password users.  Deletes the username field from registration page. Automatically makes the users chosen handle also be their display name unless Google/Twitter already has one for them.

* Fix error message display

Since the object being thrown is now an Error object (`throw new Error("uh oh!")` instead of a string `throw "uh oh!"`), we will have to check the message property of the error object in the error handler.

I also made sure this error, and the 'unkownClaimError' one, would show up and tested it out.  This way of unpacking the errors one at a time is a little verbose and hard to maintain, but that seems to be how yup does it.  I bet we could write a reducer or something to make this *just work* with new error types in the future, but this will be ok for now

---------

Co-authored-by: Steve Metz <metzsg@gmail.com>