Create new useAuthActions hook, use for Login and Register
This change fixes the name-stomping bug, and attempts to reduce complexity of the login/registration code flow.  I do this by making a new hook that will hold all app business logic for logging in and registering, with an effort to minimize code re-use and to be easy to reason about.

The hook will login, register, or link using the appropriate `firebase/auth` method, then:

For all successful authentications (login/register/link):
 - Writes uid, authProvider, and email to `/users`

Then, for registrations/links:
 - Copies the firebase User displayName to localUser.name, and
 - Writes existing LocalUser to `/users`
Else for logins:
 - Loads LocalUser from `/users`

The login and user data logic will be separate from the UI logic (showing loading indicators, error text, etc), which will reside in the Login and Register components themselves.  This will make both the hook and components more decoupled and easily testable.  (Tests to come in a follow-up).

The `Login` and `Register` components will then have a single submit handler function which simply calls the right authAction function, waits until it is done, then interprets and displays any errors, or else navigates the user to /home.

I have manually tested it with registering, linking, and logging in and things *seem* to be working nicely.  As a follow-up, I'd like to add some automated tests that mock out the firebase auth and firestore methods, so that we can run them and assure ourselves every login/registration method is fully functional, without creating tons of accounts while doing so.

I'd also like to add a loading indicator to the Login page, and a more universal indicator to Register that will appear when registering with Google/Twitter.