Use authActions when creating anon user in RoutingHelper
I missed this use case, but it should be easy to use the hook here too.

This actually helped me find a bug in useAuthActions, since it was not guaranteed to load the localUser's uid field from firestore after it was written for new registrations.  For this reason, I've adjusted the 'onSuccess' flow to call `PopulateFromFirestore` for *all* types of authentications, both login and registrations.  This will ensure a sync.  I could also refactor that code more but it would be tricky.  This is one extra read for new registrations, but this shouldn't happen too often.  We can try more optimizations in a follow-up.