Prevent logins from Register from deleting user data
Registrations with Google/Twitter actually just call `signInWithPopup`, which will succeed even if the account already exists.  In this case, there will already be a user doc.  I am opting to just log the user in to the existing, in this case, and forget that they were attempting a registration.  *An alternative would be to throw an error if registering with Google/Twitter and there is already a user doc in existence.*. Perhaps this would be preferred?