Use pure css hover effects for AnimeCard
And only use them when the device can hover (i.e. not on touchscreens).  Touchscreens can sometimes send hover states on long-press type actions but I find it is often weird and janky.

As a part of this, I also made the card slightly smaller when being hovered over.  It will help if we ever need the scrolling shelf to contain cards that expand, since they seem to always clip behind the bounds of the scroll container, so bigger means the scroll container needs a lot of padding.