Create new useAuthActions hook, use for Login and Register (#152)
* Create new useAuthActions hook, use for Login and Register

This change fixes the name-stomping bug, and attempts to reduce complexity of the login/registration code flow.  I do this by making a new hook that will hold all app business logic for logging in and registering, with an effort to minimize code re-use and to be easy to reason about.

The hook will login, register, or link using the appropriate `firebase/auth` method, then:

For all successful authentications (login/register/link):
 - Writes uid, authProvider, and email to `/users`

Then, for registrations/links:
 - Copies the firebase User displayName to localUser.name, and
 - Writes existing LocalUser to `/users`
Else for logins:
 - Loads LocalUser from `/users`

The login and user data logic will be separate from the UI logic (showing loading indicators, error text, etc), which will reside in the Login and Register components themselves.  This will make both the hook and components more decoupled and easily testable.  (Tests to come in a follow-up).

The `Login` and `Register` components will then have a single submit handler function which simply calls the right authAction function, waits until it is done, then interprets and displays any errors, or else navigates the user to /home.

I have manually tested it with registering, linking, and logging in and things *seem* to be working nicely.  As a follow-up, I'd like to add some automated tests that mock out the firebase auth and firestore methods, so that we can run them and assure ourselves every login/registration method is fully functional, without creating tons of accounts while doing so.

I'd also like to add a loading indicator to the Login page, and a more universal indicator to Register that will appear when registering with Google/Twitter.

* Use authActions when creating anon user in RoutingHelper

I missed this use case, but it should be easy to use the hook here too.

This actually helped me find a bug in useAuthActions, since it was not guaranteed to load the localUser's uid field from firestore after it was written for new registrations.  For this reason, I've adjusted the 'onSuccess' flow to call `PopulateFromFirestore` for *all* types of authentications, both login and registrations.  This will ensure a sync.  I could also refactor that code more but it would be tricky.  This is one extra read for new registrations, but this shouldn't happen too often.  We can try more optimizations in a follow-up.

* Prevent logins from Register from deleting user data

Registrations with Google/Twitter actually just call `signInWithPopup`, which will succeed even if the account already exists.  In this case, there will already be a user doc.  I am opting to just log the user in to the existing, in this case, and forget that they were attempting a registration.  *An alternative would be to throw an error if registering with Google/Twitter and there is already a user doc in existence.*. Perhaps this would be preferred?

* Bug fix: onboarding localUser was not being written to firestore.  Fix: correct typo in check of whether 'userDocAlreadyExists'.

* Delete unused 'saveToFirestore' function call when a USER has not yet been created

---------

Co-authored-by: Ken Metz <metzke@umich.edu>